api/pyproject.toml
[project]
name = "npl-api"
version = "0.1.0"
description = "Zimbabwe Cricket NPL backend API (FastAPI)"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "fastapi>=0.115.0",
    "bcrypt>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = ["httpx>=0.27.0", "pytest>=8.3.0"]

[tool.uv]
package = false
